+++ CHANGELOG.md
@@ -4,6 +4,12 @@ FaasJS use [Semantic Versioning](https://semver.org/).
 
 ## Stable
 
+[`v4.4.0 (2024-12-13)`](https://github.com/faasjs/faasjs/compare/v4.3.0...v4.4.0)
+
+- `@faasjs/server`
+  - [Feature] Add `nameMiddleware` for improve logger.
+  - [Fix] Fix staticHandler's name.
+
 [`v4.3.0 (2024-12-10)`](https://github.com/faasjs/faasjs/compare/v4.2.6...v4.3.0)
 
 - `@faasjs/server`